requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
einops==0.8.0
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.2.0
ftfy==6.3.1
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
Jinja2==3.1.3
kornia==0.7.4
kornia_rs==0.1.7
lightning-utilities==0.11.8
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==1.26.3
omegaconf==2.3.0
open_clip_torch==2.29.0
packaging==24.1
pillow==10.2.0
propcache==0.2.0
pytorch-lightning==2.4.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
sympy==1.13.1
timm==1.0.11
tokenizers==0.20.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.5.1
torchvision==0.20.1+cu121
tqdm==4.66.6
transformers==4.46.1
typing_extensions==4.9.0
urllib3==2.2.3
wcwidth==0.2.13
wget==3.2
yarl==1.17.1
